requirements.txt
streamlit
pandas
numpy
altair
psycopg2-binary
openai
python-dotenv
Flask
twilio
gunicorn
APScheduler
requests
fpdf
matplotlib
seaborn
scikit-learn
